requirements.txt
PyMuPDF>=1.24.0
google-genai>=1.39.0
langextract>=1.1.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
